requirements.txt
pyserial
autobahn
crossbar
PyQt5
idna<2.6

numpy
h5py --no-binary :all:
pyqtgraph
